renamed test_project and split it into test_helpers
